HttpParser: 测试HTTP模块

+++ EPlayerServer/Socket.h
@@ -19,7 +19,7 @@ public:
 		memcpy((char*)c_str(), str, length);
 	}
 	Buffer(const char* begin, const char* end) :std::string() {
-		int len = end - begin;
+		long int len = end - begin;
 		if (len > 0) {
 			resize(len);
 			memcpy((char*)c_str(), begin, len);
